refactor denormalizing initalizers into separate DenormalizingConcern
